refactor(front): Pass A poke gotchas — move DTOs + pure local types to types/api
Pass A handles gotcha files where a DTO references a domain type defined in the
same lib file: when that type is pure (no @app/lib coupling) it moves alongside
the DTO.

- cache, credits, workspaces: moved wholesale to types/api/poke/* (credits also
  moves the pure PokeCreditType alongside PokeUnifiedCreditRow).
- agent_configurations: partial — moved PokeAgentConfigurationType +
  PokeGetAgentConfigurationsResponseBody + PokeGetAgentDetails + PokeListSuggestions;
  kept PokeGetDatasourceRetrievalResponse in lib (references a lib observability
  type — a real coupling for Pass B).
- ~16 consumers updated; one split (poke/swr/agent_details.ts).

Deferred to Pass B (transitive @app/lib couplings): poke/conversations.ts
(ReinforcementTestCaseType -> AvailableTool), poke/global_agent_feedbacks.ts
(GlobalAgentFeedbackItem -> AgentMessageFeedbackDirection).
