+++ front-api/routes/poke/workspaces/index.ts
@@ -1,4 +1,3 @@
-import type { GetPokeWorkspacesResponseBody } from "@app/lib/api/poke/workspaces";
 import { PlanModel, SubscriptionModel } from "@app/lib/models/plan";
 import { FREE_NO_PLAN_DATA } from "@app/lib/plans/free_plans";
 import { getPlanCodeSortPriority } from "@app/lib/plans/plan_codes";
@@ -13,6 +12,7 @@ import { WorkspaceVerificationAttemptResource } from "@app/lib/resources/workspa
 import { isDomain, isEmailValid } from "@app/lib/utils";
 import { renderLightWorkspaceType } from "@app/lib/workspace";
 import logger from "@app/logger/logger";
+import type { GetPokeWorkspacesResponseBody } from "@app/types/api/poke/workspaces";
 import { normalizeError } from "@app/types/shared/utils/error_utils";
 import { pokeApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
@@ -25,7 +25,7 @@ import wId from "./[wId]";
 export type {
   GetPokeWorkspacesResponseBody,
   PokeWorkspaceType,
-} from "@app/lib/api/poke/workspaces";
+} from "@app/types/api/poke/workspaces";
 
 // Note: the parent poke/index.ts already applies pokeAuth (super-user gate).
 // This sub-router handles the workspace LIST endpoint (GET /) and mounts the
