+++ front/components/poke/credits/columns.tsx
@@ -1,8 +1,8 @@
 import { TriggerFreeCreditSegmentGrantButton } from "@app/components/poke/credits/TriggerFreeCreditSegmentGrantButton";
 import { PokeColumnSortableHeader } from "@app/components/poke/PokeColumnSortableHeader";
 import { TYPE_COLORS } from "@app/components/workspace/CreditsList";
-import type { PokeUnifiedCreditRow } from "@app/lib/api/poke/credits";
 import { getMetronomeCommitOrCreditUrl } from "@app/lib/metronome/urls";
+import type { PokeUnifiedCreditRow } from "@app/types/api/poke/credits";
 import { dateToHumanReadable } from "@app/types/shared/utils/date_utils";
 import type { WorkspaceType } from "@app/types/user";
 import { Chip, LinkWrapper } from "@dust-tt/sparkle";
