+++ front/poke/swr/agent_configurations.ts
@@ -1,6 +1,6 @@
-import type { PokeGetAgentConfigurationsResponseBody } from "@app/lib/api/poke/agent_configurations";
 import { emptyArray, useFetcher, useSWRWithDefaults } from "@app/lib/swr/swr";
 import type { PokeConditionalFetchProps } from "@app/poke/swr/types";
+import type { PokeGetAgentConfigurationsResponseBody } from "@app/types/api/poke/agent_configurations";
 import type { Fetcher } from "swr";
 
 type PokeAgentConfigurationsProps = PokeConditionalFetchProps & {
